Pass stdenv from gcc to libgcc
Pass stdenv vs relying on one supplied by callPackage. This improves the
composability of the GCC package allowing us to use a previous version
of nixpkgs' stdenv to build GCC (e.g. we want to pin nixpkgs to 22.11 to
ensure a version of GLIBC, but we want to build the latest GCC).